refactor(Experience): moving form into component, view into a component -adding selectors for both, containers, etc
