+++ src/modules/Experience/Form/ExperienceForm.styles.ts
@@ -1,6 +1,11 @@
 import { StyleSheet, ViewStyle } from 'react-native'
 
+import { Colors, colors } from '../../../styles'
+
 const styles = {
+  container: {
+    backgroundColor: colors[Colors.BackgroundGrey],
+  } as ViewStyle,
   row: {
     flexDirection: 'row',
     justifyContent: 'space-between',
