+++ src/modules/Experience/View/ExperienceView.tsx
@@ -1,0 +1,29 @@
+import { StackNavigationProp } from '@react-navigation/stack'
+import React from 'react'
+import { useTranslation } from 'react-i18next'
+
+import CvView, { CvViewCredentialTypes, CvViewList } from '../../../components/CvView'
+import CvViewCredential from '../../../components/CvView/Credential'
+import { HomeNavigationRoutes, HomeNavigatorParamsList } from '../../HomeNavigation/HomeNavigation.types'
+
+interface Props {
+  onAdd: () => void
+  navigation: StackNavigationProp<HomeNavigatorParamsList, HomeNavigationRoutes.Experience>
+  userJobs: CvViewCredentialTypes.CvViewCredentialsData
+}
+
+const ExperienceView = ({ userJobs, navigation, onAdd }: Props) => {
+  const { t } = useTranslation()
+  return (
+    <CvView
+      title={t('Experience')}
+      noDataMessage={t('Where do you currently work?')}
+      onAdd={onAdd}
+      navigation={navigation}
+    >
+      <CvViewList data={userJobs} RenderItem={CvViewCredential} />
+    </CvView>
+  )
+}
+
+export default ExperienceView
