+++ src/modules/Experience/View/index.ts
@@ -1,0 +1,3 @@
+import ExperienceView from './ExperienceView.container'
+
+export default ExperienceView
