+++ src/modules/UserJobs/index.ts
@@ -1,4 +1,4 @@
-import UserJobs from '../Experience/Experience.container'
+import UserJobs from '../Experience/View/ExperienceView.container'
 import * as middleware from './UserJobs.middleware'
 import reducer, * as actions from './UserJobs.reducer'
 import * as types from './UserJobs.types'
